chore: resume working state
Fixup the application to a working state.

The app has not been worked on over the last 10 months, and has started
to experience rot. This commit is a combination of code that wasn't
commited and enough to make the application work.

During the hiatus of development, several things have changed in the
ecosystem and will bring better development experience.
- Replacing express with fastify
- Using tap for testing
- Updating node versions