feature: timesheet review before submission
Using a stateless cookie session, we can add a step in the timesheet
submission process for the department manager to review the data before
submitting. The session cookie allows us to securely write the form
values from the backend to the cookie, rather than the database.
Preventing the need for storing of possibly temporary values.