chore: Add ObjectionJS
During development I found it difficult to work with the existing
database abstractions. An initial approach was to use @nearform/sql and
pg, but after implementing I wasn't happy with the overall structure of
the components. The next approach is was to use Objection which I've had
success with in the past.

Key reasons for going with Objection
- querying from class not only instances
- json schema support
- keep knex + migrations where needed
- easy relational modeling when needed
- graph insert/fetching