feature: Update to bootstrap 5, new theme
This was a long time coming, I never intended for the old ui to be used
more than a few weeks.

Bootstrap 5 is a tough choice in 2021, all eyes on Tailwind. I decoded
to stay with Bootstrap because it feels much more simple to work with.
Though I admit some of the utility classes take a moment to understand.

Overall this cleans up the UI nicely and prepares it to support
permission based content.